feat: improve config file error handling
